fix for blit invalid arg error
The hypothesis is that when uploading a sequence through alba_osd in
case of early failure of one of the uploads the other threads continue
to upload asynchronously, while some of the resources required for this
are already being disposed. By waiting for all threads (instead of the
early exit on failure that was there before) this situation is avoided.